requirements.txt
yarl==1.7.2
aiohttp==3.8.4
APScheduler==3.10.4
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
discord.py @ git+https://github.com/Rapptz/discord.py@879c8b9340f882b961db00de95977f3cf47c2305
distlib==0.3.5
dynaconf==3.1.12
dnspython==2.6.1
filelock==3.7.1
frozenlist==1.3.3
idna==3.3
keep==2.10.1
keep-alive==0.0.1
multidict==6.0.2
platformdirs==2.5.2
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.20.0
six==1.16.0
terminaltables==3.1.10
typing_extensions==4.3.0
virtualenv==20.15.1
wrapt==1.14.1
openai==0.27.4
mongoengine==0.27.0
pymongo==4.2.0
typing_extensions==4.3.0
pymysql==1.1.0